+++ app/src/js/index.js
@@ -20,20 +20,36 @@ document.querySelector('.js--action-url')
     });
   });
 
-  document.querySelector('.js--action-sitemap')
-    .addEventListener('click', () => {
-      const dialog = new Dialog('Enter an url of a sitemap', 'Submit Sitemap', {
-        async onSubmit(val,  input) {
-          let newSitemap;
-          try {
-            newSitemap = new Sitemap(val);
-            await newSitemap.render();
-            dialog.destroy();
-          } catch (e) {
-            console.log(e);
-            // there was an error while adding a page
-            input.value = ''; // reset input field
-          }
+document.querySelector('.js--action-sitemap')
+  .addEventListener('click', () => {
+    const dialog = new Dialog('Enter an url of a sitemap', 'Submit Sitemap', {
+      async onSubmit(val,  input) {
+        let newSitemap;
+        try {
+          newSitemap = new Sitemap(val);
+          await newSitemap.render();
+          dialog.destroy();
+        } catch (e) {
+          console.log(e);
+          // there was an error while adding a page
+          input.value = ''; // reset input field
         }
-      });
+      }
     });
+  });
+
+document.querySelector('.pane-sm .form-control').addEventListener('keyup', () => {
+  const value = document.querySelector('.pane-sm .form-control').value;
+  const items = [].slice.call(document.querySelectorAll('.list-group-item'));
+
+  if (value !== '') {
+    items.forEach((item) => {
+      const itemValue = item.querySelector('strong').textContent;
+      item.style.display = itemValue.includes(value)
+        ? 'block'
+        : 'none';
+    });
+  } else {
+    items.forEach((item) => { item.style.display = 'block'; })
+  }
+});
