+++ app/src/js/utils/clear-results.js
@@ -1,9 +1,8 @@
 const clearResults = () => {
   const results = document.querySelector('.js--result');
-  let firstChild = results.firstChild;
-  while(firstChild) {
-    firstChild.parentElement.removeChild(firstChild);
+  while(results.firstChild) {
     firstChild = results.firstChild;
+    firstChild.parentElement.removeChild(firstChild);
   }
 };
