--- js/src/index.js
@@ -1,5 +1,0 @@
-require('./online');
-require('./add-url');
-require('./add-sitemap');
-require('./search-list');
-require('./external-links');
